+++ include/ERG_Mode.h
@@ -14,6 +14,7 @@
 #define ERG_MODE_LOG_CSV_TAG "ERG_Mode_CSV"
 #define POWERTABLE_LOG_TAG   "PowTab"
 #define ERG_MODE_DELAY       700
+#define RETURN_ERROR -99
 
 extern TaskHandle_t ErgTask;
 void setupERG();